Refactored the command structure
